% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_fns.R
\name{.dt_linpred}
\alias{.dt_linpred}
\title{Internal: compute eta = X \%*\% beta with optional "accumulate" to avoid dense X}
\usage{
.dt_linpred(
  dt,
  xcols,
  beta,
  out = "eta",
  method = c("accumulate", "matrix"),
  return = c("dt", "vector")
)
}
\description{
Internal: compute eta = X \%*\% beta with optional "accumulate" to avoid dense X
}
\keyword{internal}
