% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_fns.R
\name{gee_family_fns}
\alias{gee_family_fns}
\title{Internal: get variance / mu' / linkinv for a family}
\usage{
gee_family_fns(
  family,
  link = NULL,
  dispersion = NULL,
  theta = NULL,
  precision = NULL,
  zi_prob = NULL,
  clamp_eps = 1e-08,
  varFn = NULL,
  muprimeFn = NULL,
  linkinvFn = NULL,
  ...
)
}
\description{
Internal: get variance / mu' / linkinv for a family
}
\keyword{internal}
