% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastFGEE-package.R
\docType{package}
\name{fastFGEE-package}
\alias{fastFGEE}
\alias{fastFGEE-package}
\title{fastFGEE: Fast Functional Generalized Estimating Equations}
\description{
Fits functional generalized estimating equations (fGEE) for
longitudinal functional outcomes using a one-step estimator.
}
\details{
The main user-facing functions are \code{\link{fgee}} and \code{\link{fgee.plot}}.
}
\section{Optional packages}{

\itemize{
\item \code{sanic} is used when available for faster positive-definite solves.
\item \code{irregulAR1} is used only for irregularly spaced AR1 precision matrices.
\item \code{Rcpp} and \code{RcppArmadillo} are used only for developer-side experimental
C++ loss code and are not required for the minimal CRAN build.
}
}

\section{AI-assisted development}{

Portions of the package were developed with assistance from large language
models. All code was reviewed, tested, and edited by the human author(s),
who take responsibility for the package.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gloewing/fastFGEE}
  \item Report bugs at \url{https://github.com/gloewing/fastFGEE/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Loewinger \email{gloewinger@gmail.com} (\href{https://orcid.org/0000-0002-0755-8520}{ORCID})

}
\keyword{internal}
