% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgee.plot.R
\name{fgee.plot}
\alias{fgee.plot}
\title{Default fastFGEE plotting}
\usage{
fgee.plot(
  fit,
  num_row = NULL,
  xlab = "Functional Domain",
  title_names = NULL,
  ylim = NULL,
  align_x = NULL,
  x_rescale = 1,
  y_val_lim = 1.1,
  y_scal_orig = 0.05,
  return = FALSE,
  terms.plot = TRUE,
  all.terms = TRUE,
  int.uncertainty = FALSE
)
}
\arguments{
\item{num_row}{An integer that specifies the number of rows the plots will be displayed on. Defaults to p/2, where p is the number of predictors.}

\item{xlab}{A string that specifies the x-axis title (i.e., for the functional domain). Defaults to ``Functional Domain''}

\item{title_names}{A vector of strings that has length equal to number of covariates (plus intercept if relevant). Allows one to change the titles of the plots. Defaults to NULL which uses the variable names in the dataframe for titles.}

\item{ylim}{A 2-dimensional vector that specifies limits of the y-axis in plots.}

\item{align_x}{A scalar: aligns the plot to a certain point on the functional domain and sets this as 0. This is particularly useful if the functional domain is time. Defaults to 0.}

\item{x_rescale}{A scalar: rescales the x-axis of plots which is especially useful if time is the functional domain and one wishes to, for example, account for the sampling rate. Defaults to 1.}

\item{y_val_lim}{A positive scalar that extends the y-axis by a factor for visual purposes. Defaults to 1.10. Typically does not require adjustment.}

\item{y_scal_orig}{A positive scalar that determines how much to reduce the length of the y-axis on the bottom. Defaults to 0.05. Typically does not require adjustment.}

\item{return}{Logical, indicating whether to return the data frame with the coefficient estimates and 95\% confidence intervals (CIs). Defaults to \code{FALSE}.}

\item{terms.plot}{Logical. Uses mgcv::plot.gam output to produce coefficient plots.}

\item{all.terms}{Logical. Defaults to \code{TRUE}. If set to TRUE then the partial effects of parametric model components are also plotted. See ?mgcv::plot.gam for more}

\item{int.uncertainty}{Logical. Defaults to \code{FALSE}. Uses mgcv::predict.gam() to account for uncertainty in the itercept.}

\item{fgeeobj}{A object returned from the \code{fgee} function}
}
\value{
Plots of point estimates and CIs. If \code{return = TRUE}, also returns
a list where each element is a data frame with the coefficient estimates and 95\% CIs.
}
\description{
Take a fitted \code{fgee} object produced by \code{fastFGEE::fgee()},
plots the point estimates, as well as 95\\% pointwise and joint confidence intervals.
}
\examples{
data("DTI", package = "refund")
set.seed(1)
DTI_use <- DTI[DTI$ID \%in\% sample(DTI$ID, 25),]
DTI_use <- data.frame(cca = I(DTI_use$cca),
                      case = DTI_use$case,
                      visit = 1*(DTI_use$visit),
                      sex = DTI_use$sex,
                      ID = DTI_use$ID)

fit_dti <- fastFGEE::fgee(formula = cca ~ case + visit,
                          cluster = "ID",
                          family = "gaussian", 
                          cov.type = "exchangeable"
                          data = DTI_use) # fit model
                          
fastFGEE::fgee.plot(fit_dti) # plot
}
\references{
Loewinger, G., Levis, L., Cui, E., Pereira, F. (2025). Fast Penalized
Generalized Estimating Equations for Large Longitudinal Functional Datasets. \emph{arXiv}.
}
\author{
Gabriel Loewinger \email{gloewinger@gmail.com}
}
